requirements.txt
numpy>=1.21.0
pygame>=2.1.0
PyOpenGL>=3.1.5
PyOpenGL-accelerate>=3.1.5
